refactor: give ResolvedLaunch its own module, breaking the run/sync cycle [INT-978]
The model is produced by run and consumed by sync; housing it in run
forced sync into a TYPE_CHECKING workaround. launch.py now holds the
one data model the phases hand to each other, and every import of it
is an ordinary top-level import.
